Move view and Add serializer class
