+++ hackathon_site/event/api_urls.py
@@ -7,5 +7,5 @@ app_name = "event"
 urlpatterns = [
     path("users/user/", api_views.CurrentUserAPIView.as_view(), name="current-user"),
     path("teams/team/", views.CurrentTeamAPIView.as_view(), name="current-team"),
-    path("event/leave_team/", views.LeaveTeamView.as_view(), name="leave-team"),
+    path("teams/leave_team/", api_views.LeaveTeamView.as_view(), name="leave-team"),
 ]
