+++ hackathon_site/event/api_views.py
@@ -1,7 +1,15 @@
-from rest_framework import generics, mixins
+from django.db import transaction
+from django.db.models import Q
 
-from event.models import User
-from event.serializers import UserSerializer
+from rest_framework import generics, mixins, status
+from rest_framework.exceptions import ValidationError
+from rest_framework.response import Response
+
+from event.models import User, Team as EventTeam
+from event.serializers import UserSerializer, TeamSerializer
+from event.permissions import UserHasProfile
+
+from hardware.models import OrderItem
 
 
 class CurrentUserAPIView(generics.GenericAPIView, mixins.RetrieveModelMixin):
@@ -25,3 +33,36 @@ class CurrentUserAPIView(generics.GenericAPIView, mixins.RetrieveModelMixin):
         group list are nested within the profile and user object, respectively.
         """
         return self.retrieve(request, *args, **kwargs)
+
+
+class LeaveTeamView(generics.GenericAPIView):
+    serializer_class = TeamSerializer
+    permission_classes = [UserHasProfile]
+
+    @transaction.atomic
+    def post(self, request, *args, **kwargs):
+        profile = request.user.profile
+        team = profile.team
+
+        # Raise 400 if team has active orders
+        active_orders = OrderItem.objects.filter(
+            ~Q(order__status="Cancelled"), Q(order__team=team),
+        )
+        if active_orders.exists():
+            raise ValidationError(
+                {"detail": "Cannot leave a team with already processed orders"},
+                code=status.HTTP_400_BAD_REQUEST,
+            )
+
+        # create new team
+        profile.team = EventTeam.objects.create()
+        profile.save()
+
+        # delete old team if empty
+        if not team.profiles.exists():
+            team.delete()
+
+        # Construct response data
+        response_serializer = TeamSerializer(profile.team)
+        response_data = response_serializer.data
+        return Response(data=response_data, status=status.HTTP_201_CREATED,)
